Update datepicker components to facilitate date picker average integration
